Move vars around to save memory
Its alignment requirement means that the character set can't live at our
chosen start address of $1F00.

Instead of putting a random assortment of variables in those 256 bytes,
move PRBUFF (which happens to be exactly 256 bytes) there and put all of
the other variables and the self-modifying code after the character set.

This actually increases the text buffer by 256 bytes because of the way
the start of text is calculated.  It's still 1024 bytes less than the
executable version because of the extra copy of the character set.