Run JDOS instead of the full INIT on RESET
The disk-based version of SpeedScript runs JDOS (through DOSINI) after a
system reset.  The cartridge version starts through its run vector, and
while it does detect that the program was already running, it seems to
develop some quirks after a reset.

To make the cartridge act more like the disk version, jump to JDOS when
a reset is triggered instead.

There's an interesting interaction with Altirra (also present in the
disk based version) where caps lock is turned on after a reset.  Solve
this by explicitly turning it off in the JDOS routine.