Add support for the 1200XL's F-keys
The 1200XL has F-keys with cursor movement functions, and these have
been carried over into the TK-II PS/2 keyboard adapter.  Bind the keys
as follows:

Shift-F1 (page up)	Paragraph up	SHIFT-minus
Shift-F2 (page down)	Paragraph down	SHIFT-=
Shift-F3 (home)		Top of document	START(*)
Shift-F4 (end)		End of document	CTRL-Z

The function keys are aliased to existing SpeedScript keys as shown in
the table, with TK-II key bindings in parentheses.

Shift-F3 is slightly different to the START button: it scrolls directly
to the top of the document.  START jumps to the top left of the screen,
without scrolling, the first time it is pressed.