Add support for setting a lower MEMSTART
With the right DOS, MEMLO can be much lower than $1F00.  Moving MEMSTART
to a lower address gives more room for text, but means that CHSET might
need more or less padding to be aligned to 1024 bytes.

Add tests for 512-byte and 1024-byte alignment of CHSET and pad the gap
with PRBUFF (256 bytes) and/or HDBUFF and FTBUFF (512 bytes put
together) when necessary.

The cartridge boot trick of using memory from $700-$1EFF for the text
buffer will not work if MEMSTART is below $1F00, so remove that feature
if it is.